Move workflow types into dedicated workflow type files
