+++ apps/web/app/(ee)/api/campaigns/route.ts
@@ -4,7 +4,7 @@ import { getDefaultProgramIdOrThrow } from "@/lib/api/programs/get-default-progr
 import { parseRequestBody } from "@/lib/api/utils";
 import { withWorkspace } from "@/lib/auth";
 import { prisma } from "@/lib/prisma";
-import { WorkflowAction, WorkflowCondition } from "@/lib/types";
+import { WorkflowAction, WorkflowCondition } from "@/lib/api/workflows/types";
 import {
   CampaignSchema,
   createCampaignSchema,
