+++ apps/web/lib/api/workflows/award-bounty/types.ts
@@ -1,0 +1,4 @@
+import type * as z from "zod/v4";
+import type { awardBountyConditionSchema } from "./schema";
+
+export type AwardBountyCondition = z.infer<typeof awardBountyConditionSchema>;
