+++ apps/web/lib/api/workflows/evaluate-workflow-conditions.ts
@@ -1,4 +1,4 @@
-import { WorkflowCondition, WorkflowConditionAttribute } from "@/lib/types";
+import { WorkflowCondition, WorkflowConditionAttribute } from "@/lib/api/workflows/types";
 import { COMPARISON_OPERATORS } from "./operators";
 
 export function evaluateWorkflowConditions({
