+++ apps/web/lib/api/workflows/send-campaign/types.ts
@@ -1,0 +1,5 @@
+import type * as z from "zod/v4";
+
+import type { sendCampaignConditionSchema } from "./schema";
+
+export type SendCampaignCondition = z.infer<typeof sendCampaignConditionSchema>;
