+++ apps/web/lib/bounty/api/generate-performance-bounty-name.ts
@@ -1,5 +1,5 @@
 import { isCurrencyAttribute } from "@/lib/api/workflows/utils";
-import { AwardBountyCondition } from "@/lib/types";
+import { AwardBountyCondition } from "@/lib/api/workflows/award-bounty/types";
 import { currencyFormatter, nFormatter } from "@dub/utils";
 import { PERFORMANCE_BOUNTY_SCOPE_ATTRIBUTES } from "./performance-bounty-scope-attributes";
