+++ apps/web/lib/types.ts
@@ -29,7 +29,6 @@ import {
   User,
   UtmTemplate,
   Webhook,
-  WorkflowTrigger,
   WorkspaceRole,
 } from "@prisma/client";
 import * as z from "zod/v4";
@@ -200,12 +199,6 @@ import {
   webhookEventSchemaTB,
   WebhookSchema,
 } from "./zod/schemas/webhooks";
-import {
-  WORKFLOW_ATTRIBUTES,
-  WORKFLOW_COMPARISON_OPERATORS,
-  workflowActionSchema,
-  workflowConditionSchema,
-} from "./zod/schemas/workflows";
 import { workspacePreferencesSchema } from "./zod/schemas/workspace-preferences";
 import { workspaceUserSchema } from "./zod/schemas/workspaces";
 
@@ -716,22 +709,9 @@ export type BountySubmissionRequirement =
 export type SocialMetricsChannel =
   (typeof BOUNTY_SOCIAL_PLATFORMS)[number]["value"];
 
-export type WorkflowCondition = z.infer<typeof workflowConditionSchema>;
-
-export type { AwardBountyCondition } from "./api/workflows/award-bounty/schema";
-
 export type BountySocialMetricsIncrementalBonus = z.infer<
   typeof bountySocialContentIncrementalBonusSchema
 >;
-
-export type { SendCampaignCondition } from "./api/workflows/send-campaign/schema";
-
-export type WorkflowConditionAttribute = (typeof WORKFLOW_ATTRIBUTES)[number];
-
-export type WorkflowComparisonOperator =
-  (typeof WORKFLOW_COMPARISON_OPERATORS)[number];
-
-export type WorkflowAction = z.infer<typeof workflowActionSchema>;
 
 export type BountySubmissionsQueryFilters = z.infer<
   typeof getBountySubmissionsQuerySchema
@@ -763,8 +743,6 @@ export interface TiptapNode {
   content?: TiptapNode[];
   marks?: Array<{ type: string; attrs?: Record<string, any> }>;
 }
-
-export type WorkflowAttribute = (typeof WORKFLOW_ATTRIBUTES)[number];
 
 export type EmailDomainProps = z.infer<typeof EmailDomainSchema>;
 
@@ -822,32 +800,6 @@ export type CreateFraudEventInput = Pick<
   > & {
     metadata?: Record<string, unknown> | null;
   };
-
-interface WorkflowIdentity {
-  workspaceId: string;
-  programId: string;
-  partnerId: string;
-  groupId?: string;
-  customerId?: string;
-  customerFirstSaleAt?: Date;
-}
-
-interface PartnerMetrics {
-  leads?: number;
-  conversions?: number;
-  saleAmount?: number;
-  commissions?: number;
-}
-
-export interface WorkflowContext {
-  trigger: WorkflowTrigger;
-  reason?: "lead" | "sale" | "commission";
-  identity: WorkflowIdentity;
-  metrics?: {
-    current?: PartnerMetrics;
-    aggregated?: PartnerMetrics;
-  };
-}
 
 export type SubmittedLeadProps = z.infer<typeof submittedLeadSchema>;
