+++ apps/web/lib/zod/schemas/workflows.ts
@@ -1,4 +1,3 @@
-import { WorkflowConditionAttribute } from "@/lib/types";
 import { WorkflowTrigger } from "@prisma/client";
 import * as z from "zod/v4";
 
@@ -10,6 +9,8 @@ export const WORKFLOW_ATTRIBUTES = [
   "partnerEnrolledDays",
   "partnerJoined",
 ] as const;
+
+type WorkflowConditionAttribute = (typeof WORKFLOW_ATTRIBUTES)[number];
 
 export const WORKFLOW_ATTRIBUTE_TRIGGER: Record<
   WorkflowConditionAttribute,
